Feat FindProduct : add service & implement

+++ models/web/response.go
@@ -10,6 +10,6 @@ type Pagination struct {
 	Code        int         `json:"code"`
 	Status      string      `json:"status"`
 	CurrentPage int         `json:"current_page"`
-	TotalPage   int         `json:"total_page"`
+	TotalPage   int64       `json:"total_page"`
 	Data        interface{} `json:"data"`
 }
